+++ packages/manager/src/features/EntityTransfers/EntityTransfersLanding/EntityTransfersLanding.tsx
@@ -1,19 +1,16 @@
 import { EntityTransfer } from '@linode/api-v4/lib/entity-transfers';
-import { useHistory, useLocation } from 'react-router-dom';
 import { partition } from 'ramda';
 import * as React from 'react';
+import { useHistory, useLocation } from 'react-router-dom';
 import CircleProgress from 'src/components/CircleProgress';
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import ErrorState from 'src/components/ErrorState';
 import { useEntityTransfersQuery } from 'src/queries/entityTransfers';
 import TransfersTable from '../TransfersTable';
-import ConfirmTransferDialog from './ConfirmTransferDialog';
 import CreateTransferSuccessDialog from './CreateTransferSuccessDialog';
 import TransferControls from './TransferControls';
 
 export const EntityTransfersLanding: React.FC<{}> = _ => {
-  const [confirmDialogOpen, setConfirmDialogOpen] = React.useState(false);
-  const [token, setToken] = React.useState('');
   const [successDialogOpen, setSuccessDialogOpen] = React.useState(true);
   const [transfer, setTransfer] = React.useState<EntityTransfer | undefined>(
     undefined
@@ -21,12 +18,6 @@ export const EntityTransfersLanding: React.FC<{}> = _ => {
 
   const location = useLocation();
   const history = useHistory();
-
-  const handleCloseDialog = () => {
-    setConfirmDialogOpen(false);
-    // I don't love the UX here but it seems better than leaving a token in the input
-    setTimeout(() => setToken(''), 150);
-  };
 
   const handleCloseSuccessDialog = () => {
     setSuccessDialogOpen(false);
@@ -66,16 +57,7 @@ export const EntityTransfersLanding: React.FC<{}> = _ => {
   return (
     <>
       <DocumentTitleSegment segment="Transfers" />
-      <TransferControls
-        token={token}
-        openConfirmTransferDialog={() => setConfirmDialogOpen(true)}
-        onTokenInput={setToken}
-      />
-      <ConfirmTransferDialog
-        open={confirmDialogOpen}
-        token={token}
-        onClose={handleCloseDialog}
-      />
+      <TransferControls />
       <CreateTransferSuccessDialog
         isOpen={successDialogOpen}
         transfer={transfer}
